Extract options parse
